feat: change common imports to lazy imports in AppRouter
