+++ src/pages/CvProfile/index.tsx
@@ -1,12 +1,3 @@
-import {FC} from "react";
-import {Outlet} from "react-router-dom";
-import {CvsTabs} from "@/pages/Cvs/CvsTabs";
+import {CvProfile} from "./CvProfile";
 
-export const CvProfile: FC = () => {
-  return (
-    <>
-      <CvsTabs />
-      <Outlet />
-    </>
-  );
-};
+export default CvProfile;
