+++ src/pages/Cvs/index.tsx
@@ -1,0 +1,3 @@
+import {Cvs} from "./Cvs";
+
+export default Cvs;
