+++ src/pages/Employees/index.tsx
@@ -1,0 +1,3 @@
+import {Employees} from "./Employees";
+
+export default Employees;
