+++ src/pages/Languages/index.tsx
@@ -1,0 +1,3 @@
+import {Languages} from "./Languages";
+
+export default Languages;
