+++ src/pages/Main/Main.tsx
@@ -1,11 +1,14 @@
-import {FC} from "react";
+import {FC, Suspense} from "react";
 import {Outlet} from "react-router-dom";
 import {PageLayout} from "@view/PageLayout/PageLayout";
+import {ProgressSpinner} from "@view/ProgressSpinner/ProgressSpinner";
 
 export const Main: FC = () => {
   return (
     <PageLayout>
-      <Outlet />
+      <Suspense fallback={<ProgressSpinner />}>
+        <Outlet />
+      </Suspense>
     </PageLayout>
   );
 };
