+++ src/pages/Profile/ProfileInfo/index.tsx
@@ -1,0 +1,3 @@
+import {ProfileInfo} from "./ProfileInfo";
+
+export default ProfileInfo;
