+++ src/pages/Projects/index.tsx
@@ -1,0 +1,3 @@
+import {Projects} from "./Projects";
+
+export default Projects;
